Color waves according to their type.
Shocks are red, rarefactions blue, contacts (and stationary
interface jumps) are black.  This applies to the characteristic
rays in the x-t plane as well as the hugoniot loci and integral
curves in the phase plane.

Exact Riemann solvers should return an additional list of wave types,
with one entry for each wave and each entry set to one of
{'shock' | 'raref' | 'contact'}.

For non-exact solvers, since everything is just approximated by jumps
and we don't know what kind of wave it really is, we just color
everything black.