Rewrite exposition of Traffic notebook.
This rewrite is based on the viewpoint that all the solutions with
transonic rarefactions (including those that also have a shock)
include the stationary jump as well as composite wave that is
resonant with the stationary one.

It's much clearer now why we end up with the specific solutions
that we do.